requirements.txt
flask>=3.0.0
pyyaml>=6.0.0
smbus2>=0.5.5
pyserial>=3.5
